organized the CLI files
