+++ client/lib/Animal/animalTypes.js
