+++ client/lib/Zoo/zoo.js
@@ -1,5 +1,5 @@
-const config = require('../config');
-const { getToken } = require('./tokenStore');
+const config = require('../../config');
+const { getToken } = require('../tokenStore');
 const request = require('superagent');
 const img = require('terminal-image');
